style: refactor the agents
