+++ agents/__init__.py
@@ -1,0 +1,3 @@
+from agents.mpo import MPO
+from agents.mpo_categorical import *
+from agents.mpo_humanoid import *
